# api/openapi.yml
openapi: 3.1.0
info:
  title: LibreHousehold API
  description: "This API allows interaction with the backend of LibreHousehold, an open-source household management system."
  version: 1.0.0
  contact:
    name: LibreHousehold Maintainer Nicklas
    url: https://librehousehold.wiegandt.eu
    email: nicklas+librehousehold@wiegandt.eu
servers:
  - url: https://api.librehousehold.wiegandt.eu/v1
    description: The production server for LibreHousehold API
  - url: http://localhost
    description: Local development mokapi server
  - url: http://localhost:8080/v1
    description: Local started backend server
...
paths:
...
  /household/{householdId}/tasks:
    get:
      summary: Get all tasks of a household
      description: Retrieves all tasks of the specified household. Returns an empty list if no tasks exist.
      tags:
        - tasks
      operationId: getTasks
      parameters:
        - name: householdId
          in: path
          required: true
          description: The unique identifier of the household.
          schema:
            type: string
            format: uuid
            example: "f3ce1a6b-aba1-478d-9f01-ec484f23ce60"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "404":
          description: No household found with the specified ID.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: An error occurred - see the status code and problem for more information.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      summary: Create a new task
      description: Creates a new tasks.
      tags:
        - tasks
      operationId: createTask
      parameters:
        - name: householdId
          in: path
          required: true
          description: The unique identifier of the household.
          schema:
            type: string
            format: uuid
            example: "f3ce1a6b-aba1-478d-9f01-ec484f23ce60"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: The data for the new task is invalid.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: No household found with the specified ID.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "409":
          description: The task already exists.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: An error occurred - see the status code and problem for more information.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/Problem'
  /household/{householdId}/tasks/{taskId}:
    patch:
      summary: Update a task
      description: Partially updates a task. Only the fields provided in the request body will be updated.
      tags:
        - tasks
      operationId: updateTask
      parameters:
        - name: householdId
          in: path
          required: true
          description: The unique identifier of the household.
          schema:
            type: string
            format: uuid
            example: "f3ce1a6b-aba1-478d-9f01-ec484f23ce60"
        - name: taskId
          in: path
          required: true
          description: The unique identifier of the task.
          schema:
            type: string
            format: uuid
            example: "760e8400-e29b-41d4-a716-446655440000"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: The data for updating the task is invalid.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: No household or task found with the specified ID.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: An error occurred - see the status code and problem for more information.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/Problem'
    put:
      summary: Edit a task
      description: Fully replaces all editable fields of a task.
      tags:
        - tasks
      operationId: editTask
      parameters:
        - name: householdId
          in: path
          required: true
          description: The unique identifier of the household.
          schema:
            type: string
            format: uuid
            example: "f3ce1a6b-aba1-478d-9f01-ec484f23ce60"
        - name: taskId
          in: path
          required: true
          description: The unique identifier of the task.
          schema:
            type: string
            format: uuid
            example: "760e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskEdit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: The data for editing the task is invalid.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: No household or task found with the specified ID.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: An error occurred - see the status code and problem for more information.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      summary: Delete a task
      description: Permanently deletes a task.
      tags:
        - tasks
      operationId: deleteTask
      parameters:
        - name: householdId
          in: path
          required: true
          description: The unique identifier of the household.
          schema:
            type: string
            format: uuid
            example: "f3ce1a6b-aba1-478d-9f01-ec484f23ce60"
        - name: taskId
          in: path
          required: true
          description: The unique identifier of the task.
          schema:
            type: string
            format: uuid
            example: "760e8400-e29b-41d4-a716-446655440000"
      responses:
        "204":
          description: Task successfully deleted.
        "404":
          description: No task found with the specified ID.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          description: An error occurred - see the status code and problem for more information.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
    Problem:
      #$ref: 'https://opensource.zalando.com/restful-api-guidelines/models/problem-1.0.1.yaml#/Problem'
      type: object
      properties:
        type:
          type: string
          format: uri-reference
          description: >
            A URI reference that uniquely identifies the problem type only in the
            context of the provided API. Opposed to the specification in RFC-9457,
            it is neither recommended to be dereferenceable and point to a
            human-readable documentation nor globally unique for the problem type.
          default: 'about:blank'
          example: '/some/uri-reference'
        title:
          type: string
          description: >
            A short summary of the problem type. Written in English and readable
            for engineers, usually not suited for non technical stakeholders and
            not localized.
          example: some title for the error situation
        status:
          type: integer
          format: int32
          description: >
            The HTTP status code generated by the origin server for this occurrence
            of the problem.
          minimum: 100
          maximum: 600
          exclusiveMaximum: true
        detail:
          type: string
          description: >
            A human readable explanation specific to this occurrence of the
            problem that is helpful to locate the problem and give advice on how
            to proceed. Written in English and readable for engineers, usually not
            suited for non technical stakeholders and not localized.
          example: some description for the error situation
        instance:
          type: string
          format: uri-reference
          description: >
            A URI reference that identifies the specific occurrence of the problem,
            e.g. by adding a fragment identifier or sub-path to the problem type.
            May be used to locate the root of this problem in the source code.
          example: '/some/uri-reference#specific-occurrence-context'
...
    Task:
      type: object
      required:
        - id
        - title
        - dueDate
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the task. This must be generated by the client.
        title:
          type: string
          minLength: 3
          maxLength: 200
          description: The title of the task.
        assignedTo:
          type: string
          format: uuid
          description: The id of the user the task is assigned to.
        description:
          type: string
          maxLength: 1000
          description: A detailed description of the task.
        dueDate:
          type: string
          format: date
          description: The due date of the task or if it's a recurring task, the first due date.
        recurring:
          type: boolean
          description: Indicates if the task is recurring.
        recurrenceUnit:
          type: string
          enum:
            - days
            - weeks
            - months
            - years
          description: The unit of recurrence for the task. Required if the task is recurring.
        recurrenceInterval:
          type: integer
          minimum: 1
          description: The interval of recurrence for the task. Required if the task is recurring.
        done:
          type: string
          format: date
          description: |
            The date when the task was completed. Null if the task is not yet completed.
            If the task is recurring, the task is done for the moment if the done date is after the last due date.
      example:
        id: "760e8400-e29b-41d4-a716-446655440000"
        title: "Take out the trash"
        assignedTo: "7c3a40a8-3f57-4645-8cf5-820a6a8d3613"
        description: "Take out the trash."
        dueDate: "2024-07-01"
        recurring: true
        recurrenceUnit: "weeks"
        recurrenceInterval: 1
        done: null
    TaskUpdate:
      type: object
      description: Object containing the fields to update for a task. All fields are optional.
      properties:
        done:
          type: string
          format: date
          nullable: true
          description: |
            The date when the task was completed. Set to null to mark the task as not completed.
      example:
        done: "2024-07-01"
    TaskEdit:
      type: object
      description: Object containing all editable fields of a task. Used for full replacement via PUT.
      required:
        - title
        - dueDate
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
          description: The title of the task.
          example: "Take out the trash"
        description:
          type: string
          maxLength: 1000
          description: A detailed description of the task.
          example: "Take out the trash every Monday."
        assignedTo:
          type: string
          format: uuid
          description: The id of the user the task is assigned to.
          example: "7c3a40a8-3f57-4645-8cf5-820a6a8d3613"
        dueDate:
          type: string
          format: date
          description: The due date of the task or if it's a recurring task, the first due date.
          example: "2024-07-01"
        recurring:
          type: boolean
          description: Indicates if the task is recurring.
          example: true
        recurrenceUnit:
          type: string
          enum:
            - days
            - weeks
            - months
            - years
          description: The unit of recurrence for the task. Required if the task is recurring.
          example: "weeks"
        recurrenceInterval:
          type: integer
          minimum: 1
          description: The interval of recurrence for the task. Required if the task is recurring.
          example: 1
      example:
        title: "Take out the trash"
        description: "Take out the trash every Monday."
        assignedTo: "7c3a40a8-3f57-4645-8cf5-820a6a8d3613"
        dueDate: "2024-07-01"
        recurring: true
        recurrenceUnit: "weeks"
        recurrenceInterval: 1
...